Move /accounts/<slug> to /user, enable for all
