+++ src/pages/user/index.tsx
@@ -7,157 +7,24 @@ import axios from 'axios'
 import Link from 'next/link'
 import {useViewer} from 'context/viewer-context'
 import {track} from '../../utils/analytics'
-import {Formik} from 'formik'
-import * as yup from 'yup'
-
-const emailChangeSchema = yup.object().shape({
-  email: yup.string().email().required('enter your email'),
-})
-
-export type RequestEmailChangeFormProps = {
-  originalEmail: string
-}
-
-async function requestEmailChange(newEmail: string) {
-  const {data} = await axios.post(
-    `${process.env.NEXT_PUBLIC_AUTH_DOMAIN}/api/v1/email_change_requests`,
-    {
-      email: newEmail,
-    },
-  )
-
-  return data
-}
-
-const RequestEmailChangeForm: React.FunctionComponent<RequestEmailChangeFormProps> = ({
-  originalEmail,
-}) => {
-  const VIEW_MODE = 'VIEW_MODE'
-  const EDIT_MODE = 'EDIT_MODE'
-  const [mode, setMode] = React.useState(VIEW_MODE)
-
-  const [
-    changeRequestSubmittedFor,
-    setChangeRequestSubmittedFor,
-  ] = React.useState<string | null>(null)
-
-  return (
-    <Formik
-      initialValues={{email: originalEmail}}
-      validationSchema={emailChangeSchema}
-      onSubmit={async (values) => {
-        await requestEmailChange(values.email)
-        setChangeRequestSubmittedFor(values.email)
-        setMode(VIEW_MODE)
-      }}
-    >
-      {(props) => {
-        const {
-          values,
-          isSubmitting,
-          handleChange,
-          handleBlur,
-          handleSubmit,
-          setFieldValue,
-        } = props
-        return (
-          <>
-            <form onSubmit={handleSubmit}>
-              <div className="flex flex-col space-y-2">
-                <h2 className="text-xl border-b">Email</h2>
-                <p>Your email address:</p>
-                {mode === EDIT_MODE && (
-                  <p className="text-sm text-gray-600">
-                    To ensure that you have access to this email address, we
-                    will send an email to that account with a confirmation link.
-                  </p>
-                )}
-                {changeRequestSubmittedFor && (
-                  <p className="text-sm text-gray-600">
-                    Your email change request has been received. A confirmation
-                    email has been sent to {changeRequestSubmittedFor}.
-                  </p>
-                )}
-                <div className="flex flex-row space-x-2">
-                  <input
-                    id="email"
-                    type="email"
-                    value={mode === VIEW_MODE ? originalEmail : values.email}
-                    onChange={handleChange}
-                    onBlur={handleBlur}
-                    placeholder="you@company.com"
-                    required
-                    disabled={isSubmitting || mode !== EDIT_MODE}
-                    className="bg-gray-200 focus:outline-none focus:shadow-outline border border-gray-300 rounded-md py-2 px-4 block w-full appearance-none leading-normal"
-                  />
-                  {mode === VIEW_MODE && (
-                    <button
-                      onClick={() => setMode(EDIT_MODE)}
-                      className="text-white bg-red-500 border-0 py-2 px-8 focus:outline-none hover:bg-red-600 rounded"
-                    >
-                      Edit
-                    </button>
-                  )}
-                  {mode === EDIT_MODE && (
-                    <>
-                      <button
-                        type="submit"
-                        className="text-white bg-red-500 border-0 py-2 px-8 focus:outline-none hover:bg-red-600 rounded"
-                        disabled={isSubmitting}
-                      >
-                        Submit
-                      </button>
-                      <button
-                        onClick={() => {
-                          setFieldValue('email', originalEmail)
-                          setMode(VIEW_MODE)
-                        }}
-                        className="text-black bg-gray-200 border-0 py-2 px-8 focus:outline-none hover:bg-gray-300 rounded"
-                      >
-                        Cancel
-                      </button>
-                    </>
-                  )}
-                </div>
-              </div>
-            </form>
-          </>
-        )
-      }}
-    </Formik>
-  )
-}
-
-export const getServerSideProps: GetServerSideProps = async function ({
-  req,
-  params,
-}) {
-  const {loginRequired, eggheadToken} = getTokenFromCookieHeaders(
-    req.headers.cookie as string,
-  )
-  const account: any =
-    params?.slug && (await loadAccount(params.slug as string, eggheadToken))
-  return {
-    props: {
-      loginRequired,
-      account,
-    },
-  }
-}
+import RequestEmailChangeForm from 'components/users/request-email-change-form'
+import get from 'lodash/get'
 
 type ViewerAccount = {
   stripe_customer_id: string
   slug: string
 }
 
-const Account: React.FunctionComponent<
+const User: React.FunctionComponent<
   LoginRequiredParams & {account: ViewerAccount}
-> = ({loginRequired, account = {}}) => {
+> = () => {
+  const [account, setAccount] = React.useState<any>({})
   const [subscriptionData, setSubscriptionData] = React.useState<any>()
-  const {stripe_customer_id, slug} = account
-  const {viewer} = useViewer()
+  const {stripe_customer_id} = account
+  const {viewer, authToken} = useViewer()
 
-  const {email: currentEmail} = viewer || {}
+  const {email: currentEmail, accounts} = viewer || {}
+  const {slug} = get(accounts, '[0]', {})
 
   const recur = (price: any) => {
     const {
@@ -169,6 +36,18 @@ const Account: React.FunctionComponent<
     if (interval === 'month' && interval_count === 1) return 'month'
     if (interval === 'year' && interval_count === 1) return 'year'
   }
+
+  React.useEffect(() => {
+    const loadAccountForSlug = async (slug) => {
+      if (slug) {
+        const account: any = await loadAccount(slug, authToken)
+        setAccount(account)
+      }
+    }
+
+    loadAccountForSlug(slug)
+  }, [slug, authToken])
+
   React.useEffect(() => {
     if (stripe_customer_id) {
       axios
@@ -195,8 +74,9 @@ const Account: React.FunctionComponent<
       minimumFractionDigits: 0,
     }).format(subscriptionData.price.unit_amount / 100)
 
+  // TODO: Do I need to add loginRequired prop back in below
   return (
-    <LoginRequired loginRequired={loginRequired}>
+    <LoginRequired>
       <main className="pb-10 lg:py-3 lg:px-8">
         <div className="lg:grid lg:grid-cols-12 lg:gap-x-5">
           {/* Account details */}
@@ -312,4 +192,4 @@ const Account: React.FunctionComponent<
   )
 }
 
-export default Account
+export default User
